add Product

+++ app/Http/Controllers/Admin/TagController.php
@@ -15,6 +15,7 @@ class TagController extends Controller
     protected $forUploader;
     protected $uploader;
     protected $fillField = [
+        'parent_id',
         'name_cn',
         'name_en',
         'name_jp',
@@ -34,19 +35,21 @@ class TagController extends Controller
 
     public function index()
     {
-        $tags = $this->tag->paginate();
+        $tags = $this->tag->findWithPaginate('tag');
         return view('admin.tag.index', compact('tags'));
     }
 
     public function create()
     {
-        return view('admin.tag.create');
+        $tags = $this->tag->tagList(['id', 'name_cn', 'name_en', 'name_jp']);
+        return view('admin.tag.create', compact('tags'));
     }
 
     public function store(TagRequest $request)
     {
         $path = $this->forUploader->CheckAndMarkDir('tag');
         if ($this->uploader->initialize('icon', [
+            'limit' => 1,
             'required' => true,
             'uploadDir' => $path . '/',
             'title' => 'name',
@@ -55,32 +58,57 @@ class TagController extends Controller
             'extensions' => ['jpg', 'jpeg', 'png']
         ])) {
             $file = $this->uploader->upload();
-            dd($file);
             if($warning = $this->forUploader->CheckDataWarning($file)) {
-                return back()->with('warning', $warning)->withInput($request->all());
+                return back()->with('icon', $warning)->withInput($request->all());
             }
-            $update_list = $this->forUploader->getFile($this->forUploader->CheckDataSuccess($file));
-            $this->tag->create(array_merge(array_only($request->all(), [
-
-                'icon' => base64_encode(serialize($update_list))
-            ]), $this->fillField));
+            $images = $this->forUploader->getFile($this->forUploader->CheckDataSuccess($file));
+            $this->tag->create(array_merge(array_only($request->all(), $this->fillField), [
+                'icon' => serialize($images)
+            ]));
             return redirect()->route('admin.tag.index');
         }
     }
 
     public function edit(Tag $tag)
     {
-        return view('admin.tag.edit', compact('tag'));
+        $icon = $this->forUploader->getFileType($tag->icon);
+        $tags = $this->tag->tagList(['id', 'name_cn', 'name_en', 'name_jp']);
+        return view('admin.tag.edit', compact('tag', 'icon', 'tags'));
     }
 
     public function update(TagRequest $request, Tag $tag)
     {
-        $tag->update(array_only($request->all(), $this->fillField));
-        return redirect()->route('admin.tag.index');
+        $path = $this->forUploader->CheckAndMarkDir('tag');
+        if ($this->uploader->initialize('icon', [
+            'limit' => 1,
+            'uploadDir' => $path . '/',
+            'title' => 'name',
+            'fileMaxSize' => 10,
+            'maxSize' => 50,
+            'extensions' => ['jpg', 'jpeg', 'png']
+        ])) {
+            $file = $this->uploader->upload();
+            if ($warning = $this->forUploader->CheckDataWarning($file)) {
+                return back()->with('icon', $warning)->withInput($request->all());
+            }
+            $before_file = $this->forUploader->getOldFile($this->uploader->getListInput(), $tag->icon);
+            $before_file ? $this->forUploader->crop_image($request, 'fileuploader-list-icon') : '';
+            $after_file = $this->forUploader->getFile($this->forUploader->CheckDataSuccess($file));
+            $images = $this->forUploader->checkSaveFileList($before_file, $after_file);
+            $tag->update(array_merge(array_only($request->all(), $this->fillField), [
+                'icon' => serialize($images)
+            ]));
+            return redirect()->route('admin.tag.index');
+        }
     }
 
     public function destroy(Tag $tag)
     {
+        if ($fail = $this->tag->existsChildAndProduct($tag->id)) {
+            flash($fail . 'used !')->error()->important();
+            return redirect()->route('admin.tag.index');
+        }
+        $this->forUploader->deleteFile($tag->icon);
         $tag->delete();
         return redirect()->route('admin.tag.index');
     }
